Modify community/views.py

+++ community/views.py
@@ -34,49 +34,60 @@ def article_detail(request, article_pk):
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
-        article.delete()
-        data = {
-            'delete': f'게시글 {article_pk}번이 삭제되었습니다.'
-        }
+        if request.user == article.author:
+            article.delete()
+            data = {
+                'delete': f'게시글 {article_pk}번이 삭제되었습니다.'
+            }
         return Response(data, status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        serializer = ArticleSerializer(article, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if request.user == article.author:
+            serializer = ArticleSerializer(article, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+        return Response(serializer.data)
 
 
-@api_view(['GET', 'POST'])
+@api_view(['GET'])
 @permission_classes([AllowAny])
 def comment_list(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'GET':
-        comments = get_list_or_404(Comment)
-        serializer = CommentListSerializer(comments, many=True) 
-        return Response(serializer.data)
+    comments = get_list_or_404(Comment)
+    comment_list = []
+    for comment in comments:
+        if comment.article.pk == article_pk:
+            comment_list.append(comment)
+    serializer = CommentListSerializer(comment_list, many=True) 
+    return Response(serializer.data)
     
-    elif request.method == 'POST':
-        serializer = CommentSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(article=article, author=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+@api_view(['POST'])
+@permission_classes([AllowAny])
+def comment_create(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    serializer = CommentSerializer(data=request.data)
+    if serializer.is_valid(raise_exception=True):
+        serializer.save(article=article, author=request.user)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
 @api_view(['GET', 'DELETE'])
 @permission_classes([AllowAny])
-def comment_detail(request, comment_pk):
+def comment_detail(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
     if request.method == 'GET':
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
-        comment.delete()
-        data = {
-            'delete': f'댓글 {comment_pk}번이 삭제되었습니다.'
-        }
+        if request.user == comment.author:
+            comment.delete()
+            data = {
+                'delete': f'댓글 {comment_pk}번이 삭제되었습니다.'
+            }
         return Response(data, status=status.HTTP_204_NO_CONTENT)
+
 
 
 @api_view(['POST'])
